# swagger.yaml
openapi: 3.0.0
info:
  title: LCMS API
  description: Tài liệu API cho hệ thống Quản lý Trung tâm Học tập (LCMS)
  version: 2.0.0
servers:
  - url: http://localhost:3003/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
paths:
...
  /grade-records:
    put:
      tags: [GradeRecord]
      summary: Tạo / cập nhật học bạ — TEACHER (11.1)
      operationId: upsertGradeRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertGradeRecordRequest'
      responses:
        '200':
          description: Lưu học bạ thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GradeRecord'
        '422': { $ref: '#/components/responses/ValidationError' }
  /grade-records/{id}/publish:
    patch:
      tags: [GradeRecord]
      summary: Công bố học bạ — TEACHER (11.2)
      operationId: publishGradeRecord
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Công bố thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GradeRecord'
        '404': { $ref: '#/components/responses/NotFoundError' }
  /students/{studentId}/grade-records:
    get:
      tags: [GradeRecord]
      summary: Xem học bạ học sinh (11.3)
      operationId: getStudentGradeRecords
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: string }
        - name: academic_period
          in: query
          description: "Kỳ học, ví dụ: 2025-K1"
          schema: { type: string }
      responses:
        '200':
          description: Danh sách học bạ
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/GradeRecord'

  # ── Finance (Module 12) ───────────────────────────────────────
  /invoices/vnpay/webhook:
    post:
      tags: [Finance]
      summary: VNPay IPN Webhook (12.7)
      operationId: vnpayWebhook
      description: Nhận callback từ VNPay server. Không yêu cầu JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook xử lý thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  RspCode: { type: string }
                  Message: { type: string }
  /invoices/vnpay/return:
    get:
      tags: [Finance]
      summary: VNPay Return URL (12.8)
      operationId: vnpayReturn
      description: Redirect URL sau khi người dùng hoàn tất thanh toán VNPay. Không yêu cầu JWT.
      security: []
      parameters:
        - name: vnp_ResponseCode
          in: query
          schema: { type: string }
        - name: vnp_TxnRef
          in: query
          schema: { type: string }
        - name: vnp_SecureHash
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Kết quả thanh toán
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /invoices:
    post:
      tags: [Finance]
      summary: Tạo hóa đơn thủ công — SO/BO/STAFF (12.1)
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Tạo hóa đơn thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invoice'
        '422': { $ref: '#/components/responses/ValidationError' }

    get:
      tags: [Finance]
      summary: Danh sách hóa đơn (12.3)
      operationId: getInvoices
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [unpaid, paid, overdue, cancelled]
        - name: student_id
          in: query
          schema: { type: string }
        - name: billing_period
          in: query
          description: "Kỳ học phí, ví dụ: 2025-01"
          schema: { type: string }
      responses:
        '200':
          description: Danh sách hóa đơn
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invoice'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
  /invoices/batch-generate:
    post:
      tags: [Finance]
      summary: Tạo hóa đơn hàng loạt — SO/BO (12.2)
      operationId: batchGenerateInvoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [billing_period]
              properties:
                billing_period:
                  type: string
                  description: "Kỳ học phí, ví dụ: 2025-01"
                  example: "2025-01"
                class_ids:
                  type: array
                  items: { type: string }
                  description: "Để trống = generate cho tất cả lớp active trong cơ sở"
      responses:
        '201':
          description: Tạo hóa đơn hàng loạt thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          generated: { type: integer }
                          skipped: { type: integer }
  /invoices/{id}:
    get:
      tags: [Finance]
      summary: Chi tiết hóa đơn (12.4)
      operationId: getInvoiceById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Chi tiết hóa đơn
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invoice'
        '404': { $ref: '#/components/responses/NotFoundError' }
  /invoices/{id}/pay-cash:
    post:
      tags: [Finance]
      summary: Thu tiền mặt — SO/BO/STAFF (12.5)
      operationId: payCash
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
      responses:
        '200':
          description: Thu tiền mặt thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invoice'
        '409': { $ref: '#/components/responses/ConflictError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /invoices/{id}/vnpay/create-payment:
    post:
      tags: [Finance]
      summary: Tạo link thanh toán VNPay (12.6)
      operationId: vnpayCreatePayment
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [return_url]
              properties:
                return_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Tạo link thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          payment_url:
                            type: string
                            format: uri
        '404': { $ref: '#/components/responses/NotFoundError' }

  # ── Notification (Module 13) ──────────────────────────────────
...
  /messages/threads:
    get:
      tags: [Message]
      summary: Danh sách hội thoại — TEACHER/PARENT (14.1)
      operationId: getThreads
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Danh sách hội thoại
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Thread'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
  /messages/threads/{threadId}:
    get:
      tags: [Message]
      summary: Tin nhắn trong hội thoại — TEACHER/PARENT (14.2)
      operationId: getThreadMessages
      parameters:
        - name: threadId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Tin nhắn trong hội thoại
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
  /messages:
    post:
      tags: [Message]
      summary: Gửi tin nhắn — TEACHER/PARENT (14.3)
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiver_id, content]
              properties:
                receiver_id:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Gửi tin nhắn thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
        '422': { $ref: '#/components/responses/ValidationError' }

  # ── Report (Module 15) ────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: "Thành công" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
    ApiErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Lỗi" }
        code: { type: string, example: "NOT_FOUND" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        details: { type: object }
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }

    # ── Auth ──────────────────────────────────────────────────────
...
    GradeRecord:
      type: object
      properties:
        _id: { type: string }
        studentId: { type: string }
        classId: { type: string }
        teacherId: { type: string }
        branchId: { type: string }
        academicPeriod: { type: string, example: "2025-K1" }
        assignmentAvg: { type: number, nullable: true }
        examScores:
          type: array
          items:
            type: object
            properties:
              examName: { type: string }
              score: { type: number }
              weight: { type: number }
        finalScore: { type: number, nullable: true }
        teacherComment: { type: string, nullable: true }
        status: { type: string, enum: [draft, published] }
        publishedAt: { type: string, format: date-time, nullable: true }
    UpsertGradeRecordRequest:
      type: object
      required: [student_id, class_id, academic_period]
      properties:
        student_id: { type: string }
        class_id: { type: string }
        academic_period: { type: string, example: "2025-K1" }
        exam_scores:
          type: array
          items:
            type: object
            required: [exam_name, score, weight]
            properties:
              exam_name: { type: string }
              score: { type: number, minimum: 0, maximum: 10 }
              weight: { type: number, minimum: 0, maximum: 1 }
        teacher_comment: { type: string }

    # ── Finance ───────────────────────────────────────────────────
    Invoice:
      type: object
      properties:
        _id: { type: string }
        invoiceCode: { type: string }
        branchId: { type: string }
        studentId: { type: string }
        classId: { type: string }
        billingPeriod: { type: string, example: "2025-01" }
        totalAmount: { type: number }
        status:
          type: string
          enum: [unpaid, paid, overdue, cancelled]
        paidAt: { type: string, format: date-time, nullable: true }
        paymentMethod:
          type: string
          enum: [cash, vnpay]
          nullable: true
        vnpayTransactionRef: { type: string, nullable: true }
        classSnapshot:
          type: object
          properties:
            subjectName: { type: string }
            tuitionFee: { type: number }
        createdAt: { type: string, format: date-time }
    CreateInvoiceRequest:
      type: object
      required: [student_id, class_id, billing_period, total_amount]
      properties:
        student_id: { type: string }
        class_id: { type: string }
        billing_period: { type: string, example: "2025-01" }
        total_amount: { type: number, minimum: 0 }

    # ── Notification ──────────────────────────────────────────────
...
    Thread:
      type: object
      properties:
        threadId: { type: string }
        lastContent: { type: string }
        lastSentAt: { type: string, format: date-time }
        unreadCount: { type: integer }
        otherUser:
          type: object
          properties:
            _id: { type: string }
            fullName: { type: string }
            role: { type: string }
    Message:
      type: object
      properties:
        _id: { type: string }
        threadId: { type: string }
        senderId: { type: string }
        receiverId: { type: string }
        content: { type: string }
        isRead: { type: boolean }
        sentAt: { type: string, format: date-time }

    # ── ClassAnnouncement ─────────────────────────────────────────
...
  responses:
...
    NotFoundError:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: Dữ liệu đã tồn tại hoặc xung đột trạng thái
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ValidationError:
      description: Dữ liệu gửi lên không đúng định dạng
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'